moving model_and_data to model_utils
